requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-dotenv==1.1.1
langchain==0.3.27
langgraph==0.6.2
langgraph-prebuilt==0.6.2
langchain-google-genai==2.1.8
langchain-mcp==0.2.1
langchain-mcp-adapters==0.1.9
mcp==1.12.3
